# api.yaml
openapi: 3.0.3
info:
  title: Placement Portal Application (PPA V5) API
  version: 1.0.0
  description: |
    RESTful API for a campus placement management system.
    Supports three roles: Admin, Company, and Student.
servers:
  - url: http://localhost:5001
    description: Local development server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CompanyProfile:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        company_name:
          type: string
        hr_name:
          type: string
        hr_email:
          type: string
        hr_phone:
          type: string
        website:
          type: string
        description:
          type: string
        industry:
          type: string
        company_size:
          type: string
        approval_status:
          type: string
          enum: [pending, approved, rejected]
        email:
          type: string
        is_active:
          type: boolean
        is_blacklisted:
          type: boolean
    PlacementDrive:
      type: object
      properties:
        id:
          type: integer
        company_id:
          type: integer
        company_name:
          type: string
        drive_name:
          type: string
        job_title:
          type: string
        job_description:
          type: string
        eligibility_branch:
          type: string
        min_cgpa:
          type: number
        eligible_year:
          type: integer
        application_deadline:
          type: string
          format: date-time
        location:
          type: string
        salary:
          type: string
        job_type:
          type: string
          enum: [Full-time, Internship, Contract]
        status:
          type: string
          enum: [pending, approved, rejected, closed]
        created_at:
          type: string
          format: date-time
        application_count:
          type: integer
    Application:
      type: object
      properties:
        id:
          type: integer
        student_id:
          type: integer
        drive_id:
          type: integer
        student_name:
          type: string
        student_email:
          type: string
        student_cgpa:
          type: number
        student_department:
          type: string
        drive_name:
          type: string
        company_name:
          type: string
        job_title:
          type: string
        application_date:
          type: string
          format: date-time
        status:
          type: string
          enum: [applied, shortlisted, selected, rejected, withdrawn]
        interview_type:
          type: string
        remarks:
          type: string
        cover_letter:
          type: string
    Interview:
      type: object
      properties:
        id:
          type: integer
        application_id:
          type: integer
        interview_date:
          type: string
          format: date-time
        mode:
          type: string
          enum: [Online, Offline]
        venue:
          type: string
        result:
          type: string
          enum: [pending, passed, failed]
...
    Message:
      type: object
      properties:
        message:
          type: string
paths:
...
  /api/companies/register:
    post:
      tags: [Company]
      summary: Register a new company
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [company_name, email, password]
              properties:
                company_name:
                  type: string
                email:
                  type: string
                password:
                  type: string
                hr_name:
                  type: string
                hr_phone:
                  type: string
                website:
                  type: string
                industry:
                  type: string
                description:
                  type: string
      responses:
        "201":
          description: Company registered (pending approval)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /api/companies/profile:
    get:
      tags: [Company]
      summary: Get company profile
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Company profile data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyProfile"
    put:
      tags: [Company]
      summary: Update company profile
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                company_name:
                  type: string
                hr_name:
                  type: string
                hr_email:
                  type: string
                hr_phone:
                  type: string
                website:
                  type: string
                industry:
                  type: string
                description:
                  type: string
                company_size:
                  type: string
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  company:
                    $ref: "#/components/schemas/CompanyProfile"
  /api/company/dashboard:
    get:
      tags: [Company]
      summary: Get company dashboard
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Dashboard with stats and drives summary
          content:
            application/json:
              schema:
                type: object
                properties:
                  company:
                    $ref: "#/components/schemas/CompanyProfile"
                  total_drives:
                    type: integer
                  total_applicants:
                    type: integer
                  drives_summary:
                    type: array
                    items:
                      type: object
  /api/company/drives:
    get:
      tags: [Company]
      summary: List company's drives
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of drives
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlacementDrive"
    post:
      tags: [Company]
      summary: Create a new placement drive
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [drive_name, job_title]
              properties:
                drive_name:
                  type: string
                job_title:
                  type: string
                job_description:
                  type: string
                eligibility_branch:
                  type: string
                min_cgpa:
                  type: number
                eligible_year:
                  type: integer
                application_deadline:
                  type: string
                  format: date-time
                location:
                  type: string
                salary:
                  type: string
                job_type:
                  type: string
      responses:
        "201":
          description: Drive created (pending approval)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /api/company/drives/{id}:
    put:
      tags: [Company]
      summary: Update a drive
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                drive_name:
                  type: string
                job_title:
                  type: string
                job_description:
                  type: string
                eligibility_branch:
                  type: string
                min_cgpa:
                  type: number
                eligible_year:
                  type: integer
                application_deadline:
                  type: string
                location:
                  type: string
                salary:
                  type: string
                job_type:
                  type: string
      responses:
        "200":
          description: Drive updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
    delete:
      tags: [Company]
      summary: Delete a drive
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Drive deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /api/company/drives/{id}/applications:
    get:
      tags: [Company]
      summary: Get applications for a drive
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Application"
  /api/company/drives/{id}/interviews:
    get:
      tags: [Company]
      summary: Get interviews for a drive
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of interviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Interview"
  /api/company/applications/{id}/status:
    put:
      tags: [Company]
      summary: Update application status
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [shortlisted, selected, rejected]
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /api/company/applications/{id}/schedule-interview:
    post:
      tags: [Company]
      summary: Schedule an interview for an application
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                interview_date:
                  type: string
                  format: date-time
                mode:
                  type: string
                  enum: [Online, Offline]
                venue:
                  type: string
      responses:
        "201":
          description: Interview scheduled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /api/company/applications/{id}/interview-result:
    put:
      tags: [Company]
      summary: Update interview result
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [result]
              properties:
                result:
                  type: string
                  enum: [passed, failed]
      responses:
        "200":
          description: Interview result updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

  # ─── Student ──────────────────────────────────────────
...